requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
filetype>=1.2.0
Pillow>=10.0.0
piexif>=1.1.3
geopy>=2.4.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.0.0
openpyxl>=3.1.0
extract-msg>=0.45.0
pytesseract>=0.3.10
httpx>=0.26.0
boto3>=1.34.0
structlog>=24.1.0
pyyaml>=6.0.0
redis>=5.0.0
celery>=5.3.0

# Dev / E2E
playwright>=1.40.0
pytest-playwright>=0.4.0
